# ENH Read Stata dta file incrementally