… pandas/io/stata.py
… pandas/io/tests/data/stata12_117.dta