@bashtage 2015-02-15
> I left the existing data(...) method for reading the complete file intact. Potentially it could be changed to simply call get_chunk.

It makes sense to replace this method just to avoid have quote a bit of repeated code.  Maybe 'chunksize=None` would indicate the read the entire file?