@kshedden 2015-02-18
Thanks for the review.  I think I have addressed most of the comments.  The interfaces to read_stata and StataReader should now match read_table and TextFileReader, at least in terms of features that are common to both file types.  

Comments in the current StataReader source state that it is not possible to read the value labels from the end of the dta file in a pre 117 version file without first reading through all the data.  It's not clear to me that this is the case.  Using the record size, it is possible to calculate the total size of the data in bytes, and jump from the data location to the value labels location.  I did this and all the tests pass.  However I don't know all the ins and outs of Stata files. so it is possible that I am missing something here.

I still need to work on docstrings.