@bashtage 2015-02-18
> Using the record size, it is possible to calculate the total size of the data in bytes, and jump from the data location to the value labels location. I did this and all the tests pass. However I don't know all the ins and outs of Stata files. so it is possible that I am missing something here.

I think it should be possible to seek since the binary data has a deterministic size. This was probably driven by the very old way or reading and converting element-by-element using `struct.unpack`.