@kshedden 2015-02-19
There was already a function called "_read_strls" that appears to try to read the variable length strings.  However this test suite seems to never enter this function.  I'll try to deal with this.

strls are not mentioned in the stata format 115 docs (http://www.stata.com/help.cgi?dta_115), so  I'm guessing that they were introduced in format 117.  We don't have a problem seeking to the value labels section in 117 files, the comments in the code that mentioned this limitation were explicit about it being only an issue for format <=115.

So putting this all together it looks like we might be safe jumping to the value labels based on the size of the data table (the part with fixed length records) in format <=115, and in format 117 the position of the value labels is explicitly given in the header.