@kshedden 2015-02-19
Follow-up: the strl reader seems not to work and there are checks elsewhere that raise if strls are present.  Do we still want to keep _read_strls?