@bashtage 2015-02-19
I would remove the code if it both can't be hit and is broken.