@kshedden 2015-02-19
I got read_strls working, and cleaned up the docstrings.  

I don't have anything else on my to-do list for this, so anyone who was waiting to review this more carefully can now proceed.