@jreback 2015-02-20
- can you give a run with vbench, see [here](https://github.com/pydata/pandas/wiki/Performance-Testing)
you can add ``-r stata`` to only run those benchmarks (btw, if you feel we need some more, eg. long strings / mixed strings, whatever, feel free to add). IIRC you did speed up some string stuff, so this should be nice

- pls add a release note (reference this PR number as there is not an associated issue)
- pls update the io.rst in the stata section with an example of this use (see the HDF5 or read_csv section for an example of usage).