@kshedden 2015-02-26
Does anyone know about Stata date formats like "%tdD_m_Y"?  We are not currently converting these to dates (they remain as int32) because the check for date formats needs to match "%td" exactly.  It appears to convert correctly if we treat the format as if it were simply "%td".