@kshedden 2015-02-27
It appears that %tdD_m_Y should be treated the same as %td.  The additional characters are not relevant for reading the file, and are confusing our date conversion routines.  So I am now stripping them out.