@jreback 2015-03-01
@kshedden can you squash this down to a single commit. I think ok to merge after that.